% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormdown.R
\docType{package}
\name{stormdown}
\alias{stormdown}
\title{\code{stormdown} package:}
\description{
PhD dissertation template for Harrisburg University in R Markdown
}
\details{
See the README on
\href{https://github.com/markanewman/stormdown}{GitHub}
}
